requirements.txt
mysql-connector-python
ansible
packaging